telegram_bot/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Alessandro Benetton <aleben98@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.3.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
